Music/app/requirements3.txt
Flask==2.0.1
requests==2.26.0
werkzeug==2.0.3
librosa==0.8.1
numpy== 1.19.5
tensorflow-cpu==2.6.0
pySoundFile==0.9.0.post1